Register_App/requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
google-api-core==2.19.2
protobuf>=4.25.0
requests>=2.31.0
